+++ database/migrations/2016_09_02_153301_create_likes_table.php
@@ -1,0 +1,57 @@
+<?php
+
+/*
+ * This file is part of Laravel Likeable.
+ *
+ * (c) Anton Komarev <a.komarev@cybercog.su>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+/**
+ * Class CreateLikesTable.
+ */
+class CreateLikesTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('likes', function (Blueprint $table) {
+            $table->increments('id');
+            $table->morphs('likeable');
+            $table->integer('user_id')->unsigned()->index();
+            $table->enum('type_id', [
+                'like',
+                'dislike',
+            ])->default('like');
+            $table->timestamp('created_at')->nullable();
+
+            $table->unique([
+                'likeable_id',
+                'likeable_type',
+                'user_id',
+            ], 'like_user_unique');
+
+            // $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('likes');
+    }
+}