+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to `laravel-likeable` will be documented in this file.
 
+## [4.0.0] - WIP
+
+### Changed
+
+- Database table `like` renamed to `likes`
+- Database table `like_counter` renamed to `like_counters`
+- Contracts extracted to namespace `Cog\Contracts\Likeable`
+- `ModelObserver` class renamed to `LikeableObserver`
+
 ## [3.0.0] - 2017-08-24
 
 ### Added
@@ -95,6 +104,8 @@ All notable changes to `laravel-likeable` will be documented in this file.
 
 - Initial release
 
+[4.0.0]: https://github.com/cybercog/laravel-likeable/compare/3.0.0...4.0.0
+[3.0.0]: https://github.com/cybercog/laravel-likeable/compare/2.2.5...3.0.0
 [2.2.5]: https://github.com/cybercog/laravel-likeable/compare/2.2.4...2.2.5
 [2.2.4]: https://github.com/cybercog/laravel-likeable/compare/2.2.3...2.2.4
 [2.2.3]: https://github.com/cybercog/laravel-likeable/compare/2.2.2...2.2.3